+++ Snippets.php
@@ -82,9 +82,11 @@ class SnippetsPlugin extends MantisPlugin {
 	}
 
 	public function resources($event) {
-		return '<script src="' . plugin_file("simpletip.js") . '"></script>
+		return '
 			<script src="' . plugin_file("jquery-textrange.js") . '"></script>
+			<script src="' . plugin_file("jquery.qtip.min.js") . '"></script>
 			<script src="' . plugin_file("snippets.js") . '"></script>
+			<link rel="stylesheet" type="text/css" href="' . plugin_file("jquery.qtip.min.css") . '"/>
 			<link rel="stylesheet" type="text/css" href="' . plugin_file("snippets.css") . '"/>';
 	}